添加登出逻辑，修改列表bug

+++ src/views/user/userDetail.vue
@@ -82,18 +82,25 @@ export default {
       this.serachUserData();
     },
     onDelete(data) {
-      data.isDelete = true;
-      deleteUserByPhone(data.phone)
-        .then((res) => {
-          if (res.code === 200) {
-            this.$message.success('删除成功');
-            //刷新数据
-            this.onSearch();
-          }
+      this.$confirm('确定删除吗', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+      })
+        .then(({ value }) => {
+          data.isDelete = true;
+          deleteUserByPhone(data.phone)
+            .then((res) => {
+              if (res.code === 200) {
+                this.$message.success('删除成功');
+                //刷新数据
+                this.onSearch();
+              }
+            })
+            .finally(() => {
+              data.isDelete = false;
+            });
         })
-        .finally(() => {
-          data.isDelete = false;
-        });
+        .catch(() => {});
     },
     getDateStr(time) {
       return moment(time).format('YYYY年MM月DD日 HH时:mm分:ss秒');
@@ -119,7 +126,7 @@ export default {
     serachUserData() {
       this.isLoading = true;
       let request = {
-        page: this.pageInfo.page - 1,
+        page: this.pageInfo.page ? this.pageInfo.page - 1 : 0,
         size: this.pageInfo.size,
       };
       getUserData(request)
